LMS login.md
The Student Login module is an important part of the Library Management System. It allows students to enter the system by using their correct login details. Every student receives a unique username and password when they register in the library. These login details help the system identify the correct user and provide personalized access to library services.The login page is designed to be simple and easy to use. It contains two main input fields: Username and Password. Students must type their correct information into these boxes. After entering the details, they can click the Login button to continue. If the student does not have an account, they can click the Register button to create a new account. This makes the system useful for both existing students and new users.